Move to login/components folder
