+++ 04 Login/src/pages/login/components/index.ts
@@ -1,0 +1,2 @@
+export * from './header';
+export * from './form';
