+++ 05 Recipe List/package.json
@@ -24,27 +24,27 @@
   },
   "homepage": "https://github.com/Lemoncode/vuejs-by-sample#readme",
   "devDependencies": {
-    "@types/webpack-env": "^1.13.0",
-    "awesome-typescript-loader": "^3.1.2",
-    "babel-core": "^6.24.1",
+    "@types/webpack-env": "^1.13.2",
+    "awesome-typescript-loader": "^3.4.0",
+    "babel-core": "^6.26.0",
     "babel-helper-vue-jsx-merge-props": "^2.0.2",
     "babel-plugin-syntax-jsx": "^6.18.0",
-    "babel-plugin-transform-vue-jsx": "^3.4.2",
-    "babel-preset-env": "^1.3.3",
-    "css-loader": "^0.28.0",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.11.1",
-    "html-webpack-plugin": "^2.28.0",
-    "style-loader": "^0.16.1",
-    "typescript": "^2.2.2",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.3.3",
-    "webpack-dev-server": "^2.4.2"
+    "babel-plugin-transform-vue-jsx": "^3.5.0",
+    "babel-preset-env": "^1.6.1",
+    "css-loader": "^0.28.7",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "file-loader": "^1.1.5",
+    "html-webpack-plugin": "^2.30.1",
+    "style-loader": "^0.19.0",
+    "typescript": "^2.6.1",
+    "url-loader": "^0.6.2",
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.4"
   },
   "dependencies": {
     "bootstrap": "^3.3.7",
     "lc-form-validation": "^1.0.0",
-    "vue": "^2.2.6",
-    "vue-router": "^2.4.0"
+    "vue": "^2.5.6",
+    "vue-router": "^3.0.1"
   }
 }
